learning typescript
