+++ wwwroot/board.html
@@ -1,6 +1,6 @@
 <meta charset="UTF-8">
 <link rel="stylesheet" href="./board.css" />
-<script src="./board.js"></script>
+<script type="module" src="./board.js"></script>
 
 <table id="board-id" class="chessboard"></table>
