+++ wwwroot/chess.ts
@@ -1,20 +1,14 @@
-window.onload = function() {
+import { list_games } from "./api.js";
+import { INITIAL_STATE, GAME_API } from "./config.js";
+
+window.onload = function () {
     init();
 }
 
-GAME_API = window.location.origin + '/api/game/'
-
 function init() {
     const start = document.getElementById("start-game")
-    start.addEventListener('click', newGame)
+    start!.addEventListener('click', newGame)
     populate_game_list()
-}
-
-let state = {
-    "a1": "wr", "b1": "wk", "c1": "wb", "d1": "wq", "e1": "wK", "f1": "wb", "g1": "wk", "h1": "wr",
-    "a8": "br", "b8": "bk", "c8": "bb", "d8": "bq", "e8": "bK", "f8": "bb", "g8": "bk", "h8": "br",
-    "a2": "wp", "b2": "wp", "c2": "wp", "d2": "wp", "e2": "wp", "f2": "wp", "g2": "wp", "h2": "wp", 
-    "a7": "bp", "b7": "bp", "c7": "bp", "d7": "bp", "e7": "bp", "f7": "bp", "g7": "bp", "h7": "bp"
 }
 
 function newGame() {
@@ -23,7 +17,7 @@ function newGame() {
             "content-type": "application/json"
         },
         method: 'POST',
-        body: JSON.stringify({"state": state})
+        body: JSON.stringify({ "state": INITIAL_STATE })
     })// loading animation //.then(response => window.location.href = gameUrl(response.json().id))
 }
 
@@ -36,14 +30,7 @@ function populate_game_list() {
             a.innerText = gameid
             a.href = window.location.origin + '/board.html?id=' + gameid
             li.appendChild(a)
-            lst.appendChild(li)
+            lst!.appendChild(li)
         }
     });
-}
-
-function list_games() {
-    return fetch(GAME_API)
-        .then(function(response) {
-            return response.json();
-          })
 }
