+++ wwwroot/index.html
@@ -1,5 +1,5 @@
 <meta charset="UTF-8">
-<script src="./start.js"></script>
+<script type="module" src="./chess.js" ></script>
 
 <button id="start-game" type="button"> Start a new game </button>
 <p> Or, play existing game: </p>
